move SwapAlloc to the swap module
